Make scraping more fault tolerant, only take text answers
There are some cases where there is missing data, so a question/answer cell might just be
missing. Also, we previously assumed that a daily double would not be in the lowest value
row and that is apaprently not a great assumption. Finally, .text() all of the answers,
since currently the site just renders html that's in there - they want to italicize
titles and stuff and we can ignore that.