Middleware to authenticate request is from trusted client
App and server have a shared secret that is used, along with some
portions of the url, to make it such that requests not from the mobile
app are rejected. This is mostly to keep the low-hanging fruit attacks
away - if you want to decompile my app to get the secret, then I say
you've earned it.