Adds cmd line option to seed a season from its games in db
Also add more logging - these are both for debugging purposes, my job
on ec2 randomly stopped and I didn't know why - now I have a db that
has game records with ids, but no associated jeopardy/double jeopardy.
So we should be able to populate the rest of the db with this, and
troubleshoot any problems.