Fix flaky tests causing nondeterministic Travis builds
Ah, so I converted seedFullDatabase into a module that exports stuff
so we could test parts of it. However, when you import the module
any code in it also runs, and there was some code that started seeding
the database depending on arguments in there! So sometimes it would
overwrite the database being tested midtest and lead to failures.
Pull that functionality into a function to stop it from automatically
running.