Make stop script more 'robust'
Stop and Delete both fail with error codes if the application name
isn't currently running, which is ridiculous - it fails the whole
deploy and then I have to go and start the app for no reason.

Stop and Delete all, while limiting future options (will have to be
rethought if we ever want to run more than one process), at least
only return a warning if nothing is running. So this should lead to
less failed deploys for now.